# server/routes/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Music Library Management API",
    "description": "API for managing music library integration with MusicBrainz, Lidarr, and OIDC authentication",
    "version": "1.0.0",
    "contact": {
      "name": "Music Library API Support"
    }
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "sessionAuth": []
    }
  ],
  "paths": {
...
    "/api/config/lidarr": {
      "get": {
        "tags": ["Configuration"],
        "summary": "Get Lidarr configuration",
        "description": "Get current Lidarr settings (API key masked)",
        "responses": {
          "200": {
            "description": "Lidarr configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LidarrConfig"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Configuration"],
        "summary": "Update Lidarr configuration",
        "description": "Update Lidarr connection settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "Lidarr base URL"
                  },
                  "apiKey": {
                    "type": "string",
                    "description": "Lidarr API key"
                  },
                  "rootFolder": {
                    "type": "string",
                    "description": "Root folder path"
                  },
                  "qualityProfileId": {
                    "type": "integer",
                    "description": "Quality profile ID"
                  }
                },
                "required": ["url", "apiKey", "rootFolder", "qualityProfileId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid configuration data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/lidarr/test": {
      "post": {
        "tags": ["Configuration"],
        "summary": "Test Lidarr connection",
        "description": "Test connection to Lidarr and retrieve quality profiles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "apiKey": {
                    "type": "string"
                  },
                  "useSavedApiKey": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connection test successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "version": {
                      "type": "string"
                    },
                    "profiles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Connection test failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/lidarr/rootfolders": {
      "post": {
        "tags": ["Configuration"],
        "summary": "Get Lidarr root folders",
        "description": "Retrieve available root folders from Lidarr",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "apiKey": {
                    "type": "string"
                  },
                  "useSavedApiKey": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Root folders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rootFolders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "path": {
                            "type": "string"
                          },
                          "accessible": {
                            "type": "boolean"
                          },
                          "freeSpace": {
                            "type": "integer"
                          },
                          "totalSpace": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/config/auth": {
      "get": {
        "tags": ["Configuration"],
        "summary": "Get OIDC configuration",
        "description": "Get current OIDC authentication settings (client secret masked)",
        "responses": {
          "200": {
            "description": "OIDC configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthConfig"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Configuration"],
        "summary": "Update OIDC configuration",
        "description": "Update OIDC authentication settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "issuerUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "OIDC issuer URL"
                  },
                  "clientId": {
                    "type": "string",
                    "description": "OIDC client ID"
                  },
                  "clientSecret": {
                    "type": "string",
                    "description": "OIDC client secret"
                  },
                  "domain": {
                    "type": "string",
                    "description": "Application domain"
                  }
                },
                "required": ["issuerUrl", "clientId", "clientSecret", "domain"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OIDC configuration updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "authEnabled": {
                      "type": "boolean"
                    },
                    "issuer": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid OIDC configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/auth/test": {
      "post": {
        "tags": ["Configuration"],
        "summary": "Test OIDC configuration",
        "description": "Test OIDC connection and configuration (no authentication required)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "issuerUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "clientId": {
                    "type": "string"
                  },
                  "clientSecret": {
                    "type": "string"
                  },
                  "domain": {
                    "type": "string"
                  }
                },
                "required": ["issuerUrl"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OIDC test successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "issuer": {
                      "type": "string"
                    },
                    "issuerUrl": {
                      "type": "string"
                    },
                    "discoveredEndpoints": {
                      "type": "object",
                      "properties": {
                        "authorization": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        },
                        "userinfo": {
                          "type": "string"
                        },
                        "jwks": {
                          "type": "string"
                        }
                      }
                    },
                    "clientTest": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "OIDC test failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session-based authentication using OIDC"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
      "LidarrConfig": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Lidarr base URL (without API key)"
          },
          "apiKey": {
            "type": "string",
            "description": "Masked API key (only last 4 characters shown)"
          },
          "rootFolder": {
            "type": "string",
            "description": "Root folder path for music files"
          },
          "qualityProfileId": {
            "type": "integer",
            "description": "Quality profile ID"
          }
        }
      },
      "AuthConfig": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "description": "Application domain"
          },
          "issuerUrl": {
            "type": "string",
            "description": "OIDC issuer URL"
          },
          "clientId": {
            "type": "string",
            "description": "OIDC client ID"
          },
          "clientSecret": {
            "type": "string",
            "description": "Masked client secret (only last 4 characters shown)"
          },
          "callbackUrl": {
            "type": "string",
            "description": "OIDC callback URL"
          },
          "authEnabled": {
            "type": "boolean",
            "description": "Whether authentication is enabled"
          }
        }
      }
    }
  },
...
}